Compile, then commit!
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@274360 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/optional/ssh/SSHExec.java
@@ -170,7 +170,7 @@ public class SSHExec extends SSHBase {
         try {
             // execute the command
             Session session = openSession();
-            session.setTimeout(maxwait);
+            session.setTimeout((int) maxwait);
             final ChannelExec channel=(ChannelExec) session.openChannel("exec");
             channel.setCommand(command);
             channel.setOutputStream(tee);
